% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addComments.R
\name{addComments}
\alias{addComments}
\title{addComments()}
\usage{
addComments(tre_file, ancestry_list)
}
\arguments{
\item{tre_file}{A path to your .tre output from RevBayes}

\item{ancestry_list}{A list object from createAncestry()}
}
\description{
Appends comments to an existing ancestry list from createAncestry()
}
\keyword{ancestry}
