% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAncestry.R
\name{createAncestry}
\alias{createAncestry}
\title{createAncestry()}
\usage{
createAncestry(tre_file, outputlist = TRUE, df = FALSE)
}
\arguments{
\item{tre_file}{A path to your .tre output from RevBayes}

\item{outputlist}{If you want the function to output a list}

\item{df}{If you want the function to output a dataframe}
}
\description{
Outputs a list[[]] that corresponds with the phylogenetic tree structure via library(ape).
}
\keyword{ancestry}
