% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csnap.R
\name{csnap}
\alias{csnap}
\title{csnap()}
\usage{
csnap(tre_files_path)
}
\arguments{
\item{tre_files_path}{A path to where your .tre files (assumes you have AIC, Saturated, BIC, Gene, and Uniform). Future edits will make this generalized.}
}
\description{
Create a Compare Shared Node Ages Plot (CSNAP)
}
\keyword{ancestry}
