% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csnapMulti.R
\name{csnapMulti}
\alias{csnapMulti}
\title{csnapMulti()}
\usage{
csnapMulti(tre_files_1, tre_files_2, labels_1, labels_2, study = "")
}
\arguments{
\item{tre_files_1}{A vector of 5 .tre files in comparison order}

\item{tre_files_2}{Another vector of 5 .tre files in comparison order}

\item{labels_1}{A vector of strings of labels corresponding to tre_files_1 for matrix margins}

\item{labels_2}{A vector of strings of labels corresponding to tre_files_2 for matrix margins}
}
\description{
Create a Compare Shared Node Ages Plot (CSNAP) between each pairwise sharing (AIC and BIC share these nodes / Gene and AIC share different nodes)
}
\keyword{ancestry}
