% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csnapPairs.R
\name{csnapPairs}
\alias{csnapPairs}
\title{csnapPairs()}
\usage{
csnapPairs(tre_files_path, study = "")
}
\arguments{
\item{tre_files_path}{A path to where your .tre files (assumes you have AIC, Saturated, BIC, Gene, and Uniform). Future edits will make this generalized.}
}
\description{
Create a Compare Shared Node Ages Plot (CSNAP) between each pairwise sharing (AIC and BIC share these nodes / Gene and AIC share different nodes)
}
\keyword{ancestry}
