% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNodes.R
\name{matchNodes}
\alias{matchNodes}
\title{matchNodes()}
\usage{
matchNodes(list_1, list_2)
}
\arguments{
\item{list_1}{A list object from createAncestry()}

\item{list_2}{Another list object from createAncestry()}
}
\description{
Outputs a matrix[,] that shows which nodes match on species
}
\keyword{ancestry}
