% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchNodes2.R
\name{matchNodes2}
\alias{matchNodes2}
\title{matchNodes2()}
\usage{
matchNodes2(ancestry_1, ancestry_2, intersection, num_tips)
}
\arguments{
\item{list_1}{A list object from createAncestry()}

\item{list_2}{Another list object from createAncestry()}
}
\description{
Outputs a matrix[,] that shows which nodes match on species
}
\keyword{ancestry}
